ZEN32: Decode light groups
Force configuration of the blueprint to group lights together using HA
light groups. Force the user to specify *all* the lights that the
controller covers with a single group, decode the 4 remote buttons from
the group members. (This allows the user to group several lights into a
button using a nested group.)