Testing a fix for the infinite loop.
There's an infinite loop that needs to be broken. We don't want someone
pressing on a switch, triggering a light change with this automation, and then
that light change syncing the switch, which retriggers the whole thing.

Instead, in the light-to-switch direction, break the cycle if the switch's
value has changed recently (which will be true if a human is mashing on it).